+++ ChiefMod/ChiefPlugin.cs
@@ -1,6 +1,9 @@
 using BepInEx;
+using ChiefMod.Modules;
 using HarmonyLib;
+using HunkMod.Modules.Components;
 using HunkMod.Modules.Weapons;
+using RoR2;
 
 namespace ChiefMod
 {
@@ -20,29 +23,29 @@ namespace ChiefMod
         {
             Instance = this;
             Harm = new Harmony(PluginGUID);
+            Harm.CreateClassProcessor(typeof(PickupPrefabFix)).Patch();
 
             Log.Init(Logger);
-
             ChiefSkin.Init();
-
             WeaponManager.Init();
-            WeaponManager.OnLoadCompleted += AddWeapons;
-            //ApplySoundReplacements();
+            ChiefSkin.OnChiefSkinLoaded += AddWeapons;
+
+            ApplySoundReplacements();
         }
         
-        public static void AddWeapons()
+        public static void AddWeapons(SkinDef chiefSkin)
         {
-            new PlasmaRifle().Init();
+            //new PlasmaRifle().Init();
 
-            WeaponManager.AddWeapon(SMG.instance.weaponDef, "mdlSMGHalo");
-            WeaponManager.AddWeapon(MUP.instance.weaponDef, "mdlMUPHalo");
-            WeaponManager.AddWeapon(AssaultRifle.instance.weaponDef, "mdlAssaultRifleUNSC");
-            WeaponManager.AddWeapon(Shotgun.instance.weaponDef, "mdlShotgunHalo");
+            WeaponManager.AddWeapon<SMGBehavior>(SMG.instance.weaponDef, "mdlSMGHalo", chiefSkin);
+            WeaponManager.AddWeapon<MUPBehavior>(MUP.instance.weaponDef, "mdlMUPHalo", chiefSkin);
+            WeaponManager.AddWeapon<ARBehavior>(AssaultRifle.instance.weaponDef, "mdlAssaultRifleHalo", chiefSkin);
+            WeaponManager.AddWeapon(Shotgun.instance.weaponDef, "mdlShotgunHalo", chiefSkin);
         }
 
         private void ApplySoundReplacements()
         {
-            Harm.CreateClassProcessor(typeof(EscapeSequence)).Patch();
+            //Harm.CreateClassProcessor(typeof(EscapeSequence)).Patch();
         }
     }
 }