+++ Release/CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.0.5
 
 - Hunk update
+- Merged dlls
 
 # 1.0.4